split errors into own module + more docs
