+++ src/parse.rs
@@ -114,7 +114,7 @@ impl<'p> Context<'p> {
                     MakeClosureBlock(args, None)
                 }
                 (MakeClosureBlock(args, outs), Block(code)) => {
-                    let mut inner_ctx = Context::new(code, self.source, );
+                    let mut inner_ctx = Context::new(code, self.source);
                     let code = inner_ctx.parse_block_start(cum_span.start)?;
                     let closure = Closure {
                         code,
